NET Core vs NET Std
